backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
executing==2.0.1
fastapi==0.115.12
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.7
ipykernel==6.29.5
ipython==8.27.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
nba_api==1.6.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.0.0
psycopg2==2.9.9
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.18.0
pymongo==4.10.1
PyMuPDF==1.26.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.1.49
rpds-py==0.20.0
rsa==4.9.1
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
selenium==4.25.0
Send2Trash==1.8.3
setuptools==74.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.2
supervision==0.25.1
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
tweepy==4.14.0
types-python-dateutil==2.9.0.20240821
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.49
ultralytics-thop==2.0.13
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.2
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wsproto==1.2.0
sqlalchemy==2.0.27
